requirements.txt
cracklib
psycopg2
django==1.8.1
django-celery-email==1.0.4 
https://github.com/glic3rinu/django-fluent-dashboard/archive/master.zip 
https://bitbucket.org/izi/django-admin-tools/get/a0abfffd76a0.zip 
IPy==0.81 
django-extensions==1.5.2 
django-transaction-signals==1.0.0 
django-celery==3.1.16 
celery==3.1.16 
kombu==3.0.23 
billiard==3.3.0.18 
Markdown==2.4 
djangorestframework==3.1.1
paramiko==1.15.1 
ecdsa==0.11 
Pygments==1.6 
django-filter==0.7 
passlib==1.6.2 
jsonfield==0.9.22 
lxml==3.3.5 
python-dateutil==2.2 
django-iban==0.3.0 
requests 
phonenumbers 
django-countries 
django-localflavor 
###development
django-debug-toolbar
django-nose
